Fix Configuration
