+++ src/__tests__/discourse.spec.ts
@@ -1,5 +1,5 @@
 import { DiscourseAPI } from "../Discourse";
-import { configuration } from "../Configurator";
+import { Configuration } from "../Configuration";
 
 const post =
   "\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: this is a thread\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: Another One!\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: shntnhsts\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: shntsnhts\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 a new hope\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 test post\\n\\n**archivist2**: Sorry! Something went wrong - please ask @Josh Wulf to take a look\\n\\n**Josh Wulf**: @archivist2 this is a test";
